requirements.txt
django
PyGithub
gitpython
django-q
redis
ipdb